Moved napari functionality out of reg class
